Reorder files, migrate to pyproject.toml.
