--- .circleci/config.yml
@@ -1,133 +1,0 @@
-# Python CircleCI 2.0 configuration file
-#
-# Check https://circleci.com/docs/2.0/language-python/ for more details
-#
-version: 2
-jobs:
-  build:
-    docker:
-      - image: circleci/python:3-stretch
-    working_directory: ~/repo
-
-    steps:
-      - checkout
-
-      # Download and cache dependencies
-      - restore_cache:
-          keys:
-          - v1-dependencies-{{ checksum "dev-requirements.txt" }}
-          # fallback to using the latest cache if no exact match is found
-          - v1-dependencies-
-
-      - run:
-          name: install dependencies
-          command: |
-            python3 -m venv venv
-            . venv/bin/activate
-            python -m pip install --upgrade pip
-            pip install -r dev-requirements.txt
-
-      - save_cache:
-          paths:
-            - ./venv
-          key: v1-dependencies-{{ checksum "dev-requirements.txt" }}
-        
-  test:
-    docker:
-      - image: circleci/python:3-stretch
-    working_directory: ~/repo
-
-    steps:
-      - checkout
-
-      - restore_cache:
-          key: v1-dependencies-{{ checksum "dev-requirements.txt" }}
-      - run:
-          name: set up environment
-          command: |
-            python3 -m venv venv
-            . venv/bin/activate
-            python -m pip install --upgrade pip
-            pip install -r dev-requirements.txt
-            mkdir -p /tmp/test-reports
-            mkdir -p /tmp/pytest
-            mkdir -p /tmp/htmlcov
-      - run:
-          name: pytest
-          command: venv/bin/pytest -v --cov=pymsteams --cov-report html:/tmp/htmlcov --junitxml=/tmp/pytest/results.xml >> /tmp/test-reports/pytest.out
-      - run:
-          name: safety check
-          command: |
-            grep requests setup.py | sed 's/ //g' | awk '{gsub("\x27", ""); print}' |  sed 's/,//g' | awk '{gsub("\x3E", "\x3D"); print}' | venv/bin/safety check --full-report --stdin >> /tmp/test-reports/safety.out
-      - run:
-          name: bandit analysis
-          command: venv/bin/bandit -r pymsteams >> /tmp/test-reports/bandit.out
-            
-            
-
-      - store_artifacts:
-          path: /tmp/test-reports
-          destination: test-reports
-      - store_artifacts:
-          path: /tmp/pytest
-          destination: pytest
-      - store_artifacts:
-          path: /tmp/htmlcov
-          destination: coverage
-
-      - store_test_results:
-          path: /tmp/pytest
-
-
-  deploy:
-    docker:
-      - image: circleci/python:3-stretch
-    working_directory: ~/repo
-
-    steps:
-      - checkout
-
-      - restore_cache:
-          key: v1-dependencies-{{ checksum "dev-requirements.txt" }}
-
-      - run:
-          name: verify git tag vs version
-          command: |
-            python3 -m venv venv
-            . venv/bin/activate
-            python -m pip install --upgrade pip
-            python setup.py verify
-
-      - run: 
-          name: make a source distribution
-          command: |
-            . venv/bin/activate
-            python setup.py sdist
-
-      - run:
-          name: upload to pypi
-          command: |
-            . venv/bin/activate
-            twine upload dist/*
-
-workflows:
-  version: 2
-  build_test_deploy:
-    jobs:
-      - build:
-          filters:
-            tags:
-              only: /.*/
-      - test:
-          filters:
-            tags:
-              only: /.*/
-      - deploy:
-          context: PYPI
-          requires:
-            - build
-          filters:
-            tags:
-              only: /[0-9]+(\.[0-9]+)*/
-            branches:
-              ignore: /.*/
