+++ .flake8
@@ -1,0 +1,5 @@
+[flake8]
+max-line-length = 88
+extend-ignore = E203,E501
+exclude = .git, __pycache__
+max_complexity = 10
