--- MANIFEST
@@ -1,4 +1,0 @@
-# file GENERATED by distutils, do NOT edit
-setup.cfg
-setup.py
-pymsteams/__init__.py
