--- dev-requirements.txt
@@ -1,58 +1,0 @@
-anyio>=3.2.1
-astroid>=2.4.2
-atomicwrites>=1.3.0
-attrs>=19.1.0
-bandit>=1.6.2
-bleach>=3.3.0
-certifi>=2021.5.30
-cffi>=1.14.0
-chardet>=3.0.4
-Click>=7.0
-cryptography>=3.3.2
-docutils>=0.16
-dparse>=0.5.1
-gitdb2>=2.0.5
-GitPython>=2.1.11
-h11>=0.12.0
-httpcore>=0.13.6
-httpx>=0.18.2
-idna>=2.8
-importlib-metadata>=0.18
-isort>=4.3.20
-jeepney>=0.4.3
-keyring>=21.2.1
-lazy-object-proxy>=1.4.1
-mccabe>=0.6.1
-more-itertools>=7.0.0
-packaging>=19.0
-pbr>=5.3.0
-pkginfo>=1.5.0.1
-pluggy>=0.12.0
-py>=1.10.0
-pycparser>=2.20
-Pygments>=2.7.4
-pylint>=2.5.3
-pyparsing>=2.4.0
-pytest>=5.0.1
-pytest-cov>=3.0.0
-PyYAML>=5.4
-readme-renderer>=26.0
-requests>=2.25.1
-requests-toolbelt>=0.9.1
-rfc3986>=1.5.0
-safety>=1.9.0
-SecretStorage>=3.1.2
-six>=1.12.0
-smmap2>=2.0.5
-sniffio>=1.2.0
-stevedore>=1.30.1
-toml>=0.10.1
-tqdm>=4.46.1
-twine>=3.1.1
-typed-ast>=1.4.0
-typing-extensions>=3.10.0.0
-urllib3>=1.26.5
-wcwidth>=0.1.7
-webencodings>=0.5.1
-wrapt>=1.11.2
-zipp>=0.5.1
