+++ pymsteams/exceptions.py
@@ -1,0 +1,4 @@
+class TeamsWebhookException(Exception):
+    """Custom Exception for failed webhook calls"""
+
+    pass
