+++ pyproject.toml
@@ -1,0 +1,38 @@
+[tool.poetry]
+authors = [
+    "Ryan Veach <rveach@gmail.com>",
+    "Elisiário Couto <elisiario@couto.xyz>",
+]
+name = "pymsteams"
+version = "0.2.1"
+description = "Send Adaptive Cards to Teams"
+readme = "README.md"
+license = "Apache-2.0"
+classifiers = [
+    "Development Status :: 5 - Production/Stable",
+    "Intended Audience :: Developers",
+    "Intended Audience :: Information Technology",
+    "Intended Audience :: System Administrators",
+    "Operating System :: OS Independent",
+    "Topic :: Software Development :: Libraries :: Python Modules",
+    "Topic :: Internet",
+    "Topic :: Office/Business",
+    "Topic :: Office/Business :: Groupware",
+]
+repository = "https://github.com/rveachkc/pymsteams"
+packages = [{ include = "pymsteams" }]
+
+
+[tool.poetry.dependencies]
+python = "^3.7"
+requests = "^2.28.1"
+httpx = { version = "^0.23.0", extras = ["async"] }
+
+[tool.poetry.dev-dependencies]
+pytest = "^7.1.2"
+black = "^22.6.0"
+flake8 = "^5.0.4"
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
